moves thumbnail handler class to its own file
